settings: make RegisterEnumSetting generic
We change `RegisterEnumSetting` to use any integer type instead of
always `int64`. This makes the code safer and cleaner (we no longer
need to/from `int64` casts).

Epic: none
Release note: None

+++ pkg/kv/kvserver/replicate_queue_test.go
@@ -86,7 +86,7 @@ func TestReplicateQueueRebalance(t *testing.T) {
 	for _, server := range tc.Servers {
 		st := server.ClusterSettings()
 		st.Manual.Store(true)
-		kvserver.LoadBasedRebalancingMode.Override(ctx, &st.SV, int64(kvserver.LBRebalancingOff))
+		kvserver.LoadBasedRebalancingMode.Override(ctx, &st.SV, kvserver.LBRebalancingOff)
 	}
 
 	const newRanges = 10
